Views dan Middleware

+++ database/migrations/2025_05_27_170708_create_event_register_table.php
@@ -12,29 +12,28 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('event_register', function (Blueprint $table) {
-            // === PERUBAHAN PENTING DI SINI untuk user_id ===
-            // Ganti baris ini:
-            // $table->unsignedInteger('user_id');
-            // Dengan salah satu dari dua pilihan di bawah ini:
-
-            // Pilihan 1: Ubah tipe data secara eksplisit
-            $table->unsignedBigInteger('user_id');
-            // ATAU Pilihan 2 (Direkomendasikan): Gunakan foreignId()
-            // $table->foreignId('user_id')->constrained('users')->onDelete('cascade')->onUpdate('cascade');
-
-            $table->unsignedInteger('event_id');    // Biarkan jika events.id adalah unsignedInteger
-            $table->unsignedTinyInteger('status_id'); // Biarkan jika status.id adalah unsignedTinyInteger
+            $table->unsignedInteger('user_id'); // Matches users.id
+            $table->unsignedInteger('event_id'); // Matches events.id (which is increments())
+            $table->unsignedTinyInteger('status_id'); // Matches status.id
             $table->string('payment_file')->nullable();
             $table->timestamps();
 
-            $table->primary(['user_id', 'event_id']);
+            $table->primary(['user_id', 'event_id']); // Composite primary key
 
-            // Jika menggunakan Pilihan 1 di atas (unsignedBigInteger), baris foreign key ini tetap diperlukan.
-            // Jika menggunakan Pilihan 2 (foreignId), baris foreign key untuk user_id ini bisa dihapus/dikomentari.
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade')->onUpdate('cascade');
-            
-            $table->foreign('event_id')->references('id')->on('events')->onDelete('cascade')->onUpdate('cascade');
-            $table->foreign('status_id')->references('id')->on('status')->onDelete('cascade')->onUpdate('cascade');
+            $table->foreign('user_id')
+                  ->references('id')->on('users')
+                  ->onDelete('cascade')
+                  ->onUpdate('cascade');
+
+            $table->foreign('event_id')
+                  ->references('id')->on('events')
+                  ->onDelete('cascade')
+                  ->onUpdate('cascade');
+
+            $table->foreign('status_id')
+                  ->references('id')->on('status')
+                  ->onDelete('cascade') // Or restrict/set null depending on logic
+                  ->onUpdate('cascade');
         });
     }
